% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_Tools.R
\name{Bp_Sampling}
\alias{Bp_Sampling}
\title{Ensemble of Small Models: Sampling background points using 4 different methods.}
\usage{
Bp_Sampling(
  env,
  n.points = 10000,
  method = "rand.geo",
  aggr.fact.geo = 5,
  digit.val.env = 1,
  n.strat.env = 3,
  force = FALSE,
  To.plot = FALSE,
  xy.pres = NULL
)
}
\arguments{
\item{env}{a \code{SpatRaster}. of at least one layer. if \emph{method ='rand.env'}, a minimum of 2 layers are needed.}

\item{n.points}{\code{integer}. The number of background to be selected.}

\item{method}{\code{character}. one of: "rand.geo", "strat.geo", "rand.env" or "strat.env". \emph{see Details}. \emph{Default: "rand.geo"}.}

\item{aggr.fact.geo}{\code{integer}. The aggregating factor to generate the checkerboard. Only needed 
when method = "strat.geo". \emph{see Details}. \emph{Default: 5}.}

\item{digit.val.env}{\code{integer}. The number of digit to keep to remove too similar environmental values.
Only needed when method = "rand.env". \emph{Default: 1}.}

\item{n.strat.env}{\code{integer}. The number of classes to create for each environmental layer. Only needed
when method = "strat.env". \emph{see Details}. \emph{Default: 3}.}

\item{To.plot}{\code{logical}. Should the background point plotted on a map (in black).}

\item{xy.pres}{A two-column \code{matrix} or \code{data.frame} containing the coordinates of the species occurrences.
Optional and only used when To.plot = TRUE. Plot the occurrences on the map (in aquamarine). \emph{Default: NULL}.}
}
\value{
a \code{matrix} containing the selected background points. The two first columns correspond to the coordinates.
The other columns correspond to the environmental values of these points and the last column correspond to the geographic 
or environmental classes (named "BigClass) to which the observation belongs (only for method = "strat.geo" or "strat.env")
}
\description{
This function generates background following the 4 different methods described by Steen et al (2024).
Two are generated in the environmental space whereas the remaining is in the geographic space (see details).
}
\details{
This functions generate background using the 4 methods (available with the argument method) presented in Steen et al (2024), with some adaptations to 
decrease computation time or improve the sampling strategies. Two techniques are in the geographical space and two
in the environmental space:

\enumerate{
\item 'rand.geo': Random background points selection in the geographic space. 
This is the most used techniques in SDM studies. 

\item 'strat.geo': Stratified background points selection in the geographical space. To 
realize this stratification, a checkerboard is created with a pixel of a certain resolution. The size of these
pixels can be modified with the argument 'aggr.fact.geo'. For each pixel of the checkerboard the same number of 
background points are randomly selected to reach the value of \emph{n.points}. 

\item  'rand.env': Random background point selection in the environmental space. 
To do so, a PCA is first performed and the two fist axes are kept to 
reflect the environmental space. To reduce too similar points, the PCA scores of each observation are rounded 
at a certain digit (argument \emph{digit.val.env}) and only one observation among the similar ones is kept.
Note that increasing the value of digit.val.env will strongly increase computation time.
A convex hull is then generated to encompass all environmental points in the PCA. This polygon is split into 
\emph{n} smaller polygons of roughly the same size (derived from the argument \emph{n.points})
One observation is then capture in each of these small polygons. A second random selection is afterwards performed 
in the small polygons that still have at least one observation to reach the requested number of background points. 
By doing this, the entire environmental is captured avoiding the over-abundance of some common environment. 

\item 'strat.env': Stratified background points selection in the environmental space.
To perform this selection, each environmental layer is converted into classes of \emph{n.strat.env}
by dividing the range of the environmental predictor by \emph{n.strat.env} of the same size. For each pixel 
constituting the grid, we then combined all the predictor together, creating a combination of classes. For each 
of the different possible combination of classes available, the same number of background point is selected. if
the needed number of background points is not available for a certain category, all the observations from this 
category are taken and the remaining needed background points are collected in the other categories to reach 
the value of \emph{n.points}.

}
}
\examples{
\donttest{
library(terra)
env <- terra::unwrap(ESM_Env)
# Selection full random in the geographic space
Bp <- Bp_Sampling(env = env,
                  n.points = 1000,
                  method = "rand.geo",
                  To.plot = FALSE)
                       
# Selection stratified  in the geographic space                     
Bp <- Bp_Sampling(env = env,
                  n.points = 1000,
                  method = "strat.geo",
                  aggr.fact.geo = 100,
                  To.plot = FALSE)
                       
# Selection full random in the environmental space                     
#  Bp <- Bp_Sampling(env = env,
#                    n.points = 1000,
#                    method = "rand.env",
#                    digit.val.env = 2,
#                    To.plot = FALSE)       
                                        
# Selection stratified in the environmental space                     
Bp <- Bp_Sampling(env = env,
                  n.points = 1000,
                  method = "strat.env",
                  n.strat.env = 3,
                  To.plot = FALSE)       
}               
}
\references{
Steen,B., Broennimann, O., Maiorano, L., Guisan, . 2024. How sensitive are species distribution models to different background point 
selection strategies? A test with species at various equilibrium levels. 
\emph{Ecological Modelling}. \bold{493}, 110754. \doi{10.1016/j.ecolmodel.2024.110754}.
}
\seealso{
\code{\link{ESM_Modeling}}
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com}
}
