\name{dmat} 
\Rdversion{1.1} 
\alias{dmat} 
\title{Discretisation of an interaction matrix}
\description{Returns an interaction matrix with n performance categories}
\usage{ 
dmat(m,n=2)
}

\arguments{ 
\item{m}{An interaction matrix} 
\item{n}{Number of categories} 
}

\value{
	\item{}{A dicretised interaction matrix}
}