% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_Tools.R
\name{ESM_Binarize}
\alias{ESM_Binarize}
\title{Binarize probability values using a threshold}
\usage{
ESM_Binarize(proj, thr)
}
\arguments{
\item{thr}{\code{numeric}. threshold to binarize the probabilities. \bold{must be a single value}.}
}
\value{
An object of the same class as proj
}
\description{
This function binarizes probability values based on a specific threshold
}
\details{
proj < thr will return 0 while proj >= thr will be equal to 1.
}
\examples{
## Generate a vector to binarize
proj <- seq(0,1000, by = 100)
# Binraization of the vector
ESM_Binarize(proj = proj, thr = 400)

}
\seealso{
[ESM_Projection], [ESM_Ensemble.Projection], [ESM_Threshold]
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com}
}
