% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_BpSampling.R
\name{ESM_Bp.Sampling}
\alias{ESM_Bp.Sampling}
\title{Ensemble of Small Models: Sampling background points using 4 different methods.}
\usage{
ESM_Bp.Sampling(
  env,
  n.points = 10000,
  method = "rand.geo",
  aggr.fact.geo = 5,
  n.strat.env = 3,
  round.val.env = 2,
  To.plot = FALSE,
  xy.pres = NULL
)
}
\description{
Ensemble of Small Models: Sampling background points using 4 different methods.
}
\references{
Steen,B., Broennimann, O., Maiorano, L., Guisan, . 2024. How sensitive are species distribution models to different background point 
selection strategies? A test with species at various equilibrium levels. 
\emph{Ecological Modelling}. \bold{493}, 110754. \doi{10.1016/j.ecolmodel.2024.110754}.
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com}
}
