% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_EnsembleModeling.R
\name{ESM_Ensemble.Modeling}
\alias{ESM_Ensemble.Modeling}
\title{Ensemble of Small Models: Average Bivariate Models into an ESM}
\usage{
ESM_Ensemble.Modeling(
  ESM.Mod,
  weighting.score,
  threshold = NULL,
  save.obj = TRUE
)
}
\arguments{
\item{ESM.Mod}{The object returned by \code{ESM_Modeling}.}

\item{weighting.score}{\code{character}. An evaluation score used to weight single models to build ensembles:'AUC','MaxTSS', 'SomersD', or '
SBI' (if SBI = TRUE in ESM_Modeling) or 'Boyce' (if FALSE).}

\item{threshold}{\code{numeric} or \code{NULL}. Threshold value of an evaluation score to select the bivariate model(s) included for building 
the ensemble. \emph{Default}: 0.5 for AUC and 0 for the other metrics.}

\item{save.obj}{\code{logical}. Allows or not to save the outputs from this function.}
}
\value{
\itemize{
a \code{list} containing: 
\item{data}: a \code{list} with the object resp, xy, env.var and sp.name. env.var is = to the data supplied in the argument env. 
If env, was a SpatRaster, it corresponds to the extracted values of these rasters for the species coordinates.
\item{model.info}: a \code{list} of the models used, their options, the combination of bivariate models (which.biva), the modeling ID, 
the path to the folder where are the stored the models (biva.path), and the failed models
\item{cv.split.table}: a \code{matrix} used to train and test models. See explanation of the argument cv.split.table
\item{evaluations}: a \code{matrix} The evaluation of the ensemble based on 4 metrics: the AUC, the Somer's D (=2*AUC-1),
maxTSS, and the smooth Boyce Index (SBI, if SBI = TRUE in ESM_Modeling) or the regular Boyce Index (if SBI = FALSE).
\item{EF.algo}: a \code{list} containing \code{pred.EF.algo} which is a \code{matrix} of ensemble predictions for each run 
and modeling techniques; and \code{weights.algo} a \code{matrix} of weights used to generate the ensemble at the level of the algorithm.
\item{EF}: a \code{list} containing \code{pred.EF} which is a \code{matrix} of ESMs predictions for each run; 
and \code{weights.EF} a \code{matrix} of weights used to generate the ESMs. Note that if only one modelling technique is used 
\code{EF} will be exactly the same as \code{EF.algo}.
}
}
\description{
This function averages simple bivariate models by weighted means to Ensemble Small Models.
}
\details{
For the use of this function, please refer to the manual of ESM_Modeling.
}
\references{
Lomba, A., L. Pellissier, C.F. Randin, J. Vicente, F. Moreira, J. Honrado and A. Guisan. 2010. Overcoming the rare species 
modelling paradox: A novel hierarchical framework applied to an Iberian endemic plant. \emph{Biological Conservation}, \bold{143},2647-2657.

Breiner F.T., A. Guisan, A. Bergamini and M.P. Nobis. 2015. Overcoming limitations of modelling rare species by using ensembles of small models. \emph{Methods in Ecology and Evolution}, \bold{6},1210-1218.

Breiner F.T., Nobis M.P., Bergamini A., Guisan A. 2018. Optimizing ensembles of small models for predicting the distribution of species with few occurrences. \emph{Methods in Ecology and Evolution}. \doi{10.1111/2041-210X.12957}
}
\seealso{
\code{\link{ESM_Modeling}}, \code{\link{ESM_Projection}} and  \code{\link{ESM_Ensemble.Projection}}
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com}
}
