% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_EnsembleProjection.R
\name{ESM_Ensemble.Projection}
\alias{ESM_Ensemble.Projection}
\title{Ensemble of Small Models: Projections of ESMs}
\usage{
ESM_Ensemble.Projection(
  ESM.proj,
  ESM.ensembleMod,
  chosen.models = "all",
  save.obj = TRUE
)
}
\arguments{
\item{ESM.ensembleMod}{The object returned by \code{\link{ESM_Ensemble.Modeling}}.}

\item{chosen.models}{a \code{character} allowing to only make the ensemble maps on a selection of modeling techniques. \emph{Default}: "all".}

\item{save.obj}{\code{logical}. If \code{TRUE}, the map or the data.frame will be saved. Maps or data.frame wiil be stored in the 
folder named by sp.name. Values are multiplied by 1000 and rounded. Note that the maps are in .tif and compressed with the option 
COMPRESS=DEFLATE and PREDICTOR=2.}

\item{ESM.Mod}{The object returned by \code{\link{ESM_Projection}}.}
}
\value{
a \code{SpatRaster} or \code{data.frame}. EF = the ensemble across the modeling techniques.
}
\description{
Generate the ESMs into a new space
}
\references{
Lomba, A., L. Pellissier, C.F. Randin, J. Vicente, F. Moreira, J. Honrado and A. Guisan. 2010. Overcoming the rare species 
modelling paradox: A novel hierarchical framework applied to an Iberian endemic plant. \emph{Biological Conservation}, \bold{143},2647-2657.

Breiner F.T., A. Guisan, A. Bergamini and M.P. Nobis. 2015. Overcoming limitations of modelling rare species by using ensembles of small models. \emph{Methods in Ecology and Evolution}, \bold{6},1210-1218.

Breiner F.T., Nobis M.P., Bergamini A., Guisan A. 2018. Optimizing ensembles of small models for predicting the distribution of species with few occurrences. \emph{Methods in Ecology and Evolution}. \doi{10.1111/2041-210X.12957}
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com} based on the previous code written by Frank Breiner 
and Mirko Di Febbraro with the contributions of Olivier Broennimann and Flavien Collart
}
