% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ESM_Env}
\alias{ESM_Env}
\title{Environmental SpatRaster to perform ESM with the ESM package}
\format{
A SpatRaster with 4 layers:
\describe{
\item{ch_edaphic_eivdescombes_pixel_r}{Ecological indicator values reflecting 
Soil pH.}
\item{sradY}{Annual solar radiation. Sum of monthly solar radiations}
\item{bio3_tiso_8110_LV95}{Isothermalithy. It is equal to: Mean Diurnal 
Range/Temperature Annual Range * 100. The values have been averaged between 
1981 and 2010}
\item{bio15_ps_8110_LV95}{Precipitation Seasonality. Computed using the 
coefficient of variation among monthly precipitation and averaged between
1981-2010.}
}
}
\usage{
ESM_Env
}
\description{
Contains stack of 4 predictors at 25m resolution in the Western Swiss Alps.
}
\examples{
library(terra)
data(ESM_Env)
ESM_Env <- terra::unwrap(ESM_Env)

}
\references{
Collart, F., Kiebacher, T., Quetsch, M., Broennimann, O., Guisan, A. & Vanderpoorten, A. 2024. To what extent can we predict 
variation of bryophyte and tracheophyte community composition at fine spatial scale along an elevation gradient?
\emph{STOTEN}, 926:171741. \doi{10.1016/j.scitotenv.2024.171741}.

Külling, N., Adde, A., Fopp, F. et al. 2024. SWECO25: a cross-thematic raster 
database for ecological research in Switzerland. \emph{Sci Data}, 11:21. 
\doi{10.1038/s41597-023-02899-1}.
}
\author{
Flavien Collart
}
\keyword{datasets}
