% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_NullModels.R
\name{ESM_Null.Models}
\alias{ESM_Null.Models}
\title{Ensemble of Small Models: Evaluation using Null Models}
\usage{
ESM_Null.Models(
  ESM.Mod,
  ESM.ensembleMod,
  n.rep = 99,
  quant = 95,
  pooling = FALSE,
  hist.plot = FALSE,
  parallel = FALSE,
  n.cores = 1,
  pathToSaveObject = getwd(),
  save.obj = FALSE
)
}
\arguments{
\item{ESM.Mod}{The object returned by \code{ESM_Modeling}.}

\item{ESM.ensembleMod}{The object returned by \code{\link{ESM_Ensemble.Modeling}}.}

\item{n.rep}{\code{integer}. The number of null models. \emph{Default: 99}.}

\item{quant}{\code{numeric}. The quantile used to recalibrated the evaluation metrics (between 0 and 1). \emph{Default: 0.95}.}

\item{pooling}{\code{logical}. Should the pooling evaluation be performed? \emph{Default: FALSE}.}

\item{hist.plot}{\code{logical}. Should an histogram representing the null distribution be performed? \emph{Default: FALSE}.}

\item{parallel}{\code{logical}. Allows or not parallel job using the function parallel::makeCluster. \emph{Default: FALSE} 
but It is highly recommended to perform the null models in parallel as it takes quite a lot of time.}

\item{n.cores}{\code{integer}. Number of cores used to make the models.}

\item{pathToSaveObject}{The path to save the objects (temporally if save.obj = FALSE). \emph{Default: working directory}}

\item{save.obj}{\code{logical}. Should all the null models be kept? \emph{Default: FALSE}.}
}
\value{
\itemize{
a \code{list} containing: 
\item{pval}: \code{numeric}. The pvalue computed for each metric.
\item{adj.evaluation}: \code{numeric}. Adjusted evaluation metrics for the ESM based on a certain quantile from the null distribution.
\item{evaluations}: \code{matrix}. Observed and null values for each metrics.
}
}
\description{
Performed null models to test the significativity of the ESM evaluatuion as recommended in Collart & Guisan (2023) and performed in 
van Proosdij et al (2016). 
It consists of running a series of null models, randomly shuffling the presences and absences, thus generating ‘fake’ occurrences.
These models are performed folowing the same methodology as for the ESM (same model parameters, same number of cross-validations and 
same threshold for the ensemble). The pooling evaluation can also be combined with these null models with the option pooling = TRUE 
(as recommended in Collart & Guisan, 2023). In addition to the pvalue, the observed evaluation metrics are then reajusted using the 
value of the evaluation metric at a certain quantile following the methology developped in Verdon et al (2024).
}
\references{
Collart, F., & Guisan, A. 2023. Small to train, small to test: Dealing with low sample size in model evaluation. 
\emph{Ecological Informatics}. \bold{75}, 102106. \doi{10.1016/j.ecoinf.2023.102106}.

van Proosdij, A.S.J., Sosef, M.S.M., Wieringa, J.J. and Raes, N. 2016. Minimum required number of specimen records to 
develop accurate species distribution models. \emph{Ecography}. \bold{39}, 542-552. \doi{10.1111/ecog.01509}.

Verdon, V., Malard, L., Collart, F., Adde, A., Yashiro, E., Lara Pandi, E., Mod, H., Singer, D., Niculita-Hirzel, H., 
Guex, N. and Guisan, A. 2024. Can we accurately predict the distribution of soil microorganism presence and relative abundance? 
\emph{Ecography}. e07086. \doi{10.1111/ecog.07086}.
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com}
}
