% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_Projection.R
\name{ESM_Projection}
\alias{ESM_Projection}
\title{Ensemble of Small Models: Projections of Bivariate Models}
\usage{
ESM_Projection(
  ESM.Mod,
  new.env,
  name.env,
  rounded = TRUE,
  pred.multiplier = 1000,
  parallel = FALSE,
  n.cores = 1,
  datatype = "INT2U",
  save.obj = T,
  verbose = TRUE
)
}
\arguments{
\item{ESM.Mod}{The object returned by \code{\link{ESM_Modeling}}.}

\item{new.env}{a \code{data.frame}, \code{matrix} or a \code{SpatRaster} of the predictors where the models 
should be projected. Note that the colnames or the names of the predictors should be exactly the same as the one used in \code{\link{ESM_Modeling}}.}

\item{name.env}{a \code{character} which will be used to generate a folder to save the individual projections.}

\item{rounded}{\code{logical}. Should the prediction be rounded to become an integer? \emph{Default:} TRUE.}

\item{pred.multiplier}{\code{numeric}. The factor to multiply the predictions before rounding them (must be a power of 10). 
Only needed when rounded = FALSE. \emph{Default:} 1000.}

\item{parallel}{\code{logical}. Allows or not parallel job using the functions makeCluster.}

\item{n.cores}{\code{integer}. Number of CPU cores used to make the models.}

\item{datatype}{a \code{character} When new.env is a  \code{SpatRaster}, which datatype should be used to save the files? \emph{Default:} 'INT2U' if rounded = TRUE 
and 'FLT4S' if rounded = FALSE.}

\item{save.obj}{\code{logical}. Allows or not to save the final output into a '.out file'.}

\item{verbose}{\code{logical}. Allows or not message.}
}
\value{
a \code{list} containing: 
\itemize{
\item{projection.path}: a \code{character} the path to each projection.
\item{model.info}: a \code{list} of the models used, their options, the combination of bivariate models (which.biva),
the modeling ID, the path to the folder where are the stored the models (biva.path), and the failed models.
\item{name.env}: a \code{character} the value given to the argument name.env.
\item{proj.type}: a \code{character} which is either "data.frame" or "SpatRaster" depending on the class of new.env
}
}
\description{
Project bivariate models
}
\details{
\describe{
Each bivariate models are geographically projected into the new.env area. 
Note that if new.env is a SpatRaster, the projected maps are by default multiplied by 1000 and 
rounded to reduce storage space. They are also store in compress tif file (with these parameters: "COMPRESS=DEFLATE","PREDICTOR=2").
If the argument 'rounded' is set to FALSE and datatype is left default, the datatype will be changed to 'FLT4S' to keep decimals.
For the use of this function, please refer to the manual of ESM_Modeling.
}
}
\references{
Lomba, A., L. Pellissier, C.F. Randin, J. Vicente, F. Moreira, J. Honrado and A. Guisan. 2010. Overcoming the rare species 
modelling paradox: A novel hierarchical framework applied to an Iberian endemic plant. \emph{Biological Conservation}, \bold{143},2647-2657.

Breiner F.T., A. Guisan, A. Bergamini and M.P. Nobis. 2015. Overcoming limitations of modelling rare species by using ensembles of small models. \emph{Methods in Ecology and Evolution}, \bold{6},1210-1218.

Breiner F.T., Nobis M.P., Bergamini A., Guisan A. 2018. Optimizing ensembles of small models for predicting the distribution of species with few occurrences. \emph{Methods in Ecology and Evolution}. \doi{10.1111/2041-210X.12957}
}
\seealso{
\code{\link{ESM_Modeling}}, \code{\link{ESM_Ensemble.Modeling}}, \code{\link{ESM_Ensemble.Projection}}
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com}
}
