% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_Tools.R
\name{ESM_Range.Shift.Continuous}
\alias{ESM_Range.Shift.Continuous}
\title{Ensemble of Small Models: Range Shift between projections with continuous predictions}
\usage{
ESM_Range.Shift.Continuous(
  proj.curr,
  proj.fut,
  occ.pres = NULL,
  test.significance = TRUE
)
}
\arguments{
\item{proj.curr}{\code{SpatRaster}. One ESM/SDM continuous projections at current time.}

\item{proj.fut}{\code{SpatRaster}. One or more ESM/SDM continuous projections at future time.}
}
\description{
This function compare ESM or SDM continuous projections between two time periods by XXXX
}
\details{
The comparisons between SDM/ESM continuous predictions are made following this formula:
}
\seealso{
\code{\link{ESM_Projection}}, \code{\link{ESM_Ensemble.Projection}}
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com}
}
