% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_Tools.R
\name{ESM_Range.Shift}
\alias{ESM_Range.Shift}
\title{Ensemble of Small Models: Range Shift between projections}
\usage{
ESM_Range.Shift(proj.curr, proj.fut)
}
\arguments{
\item{proj.curr}{\code{SpatRaster}. One or more ESM/SDM binary projections at current time. 
\emph{Note that if the number of layer is greater than one for proj.curr, the number of layer 
in proj.fut must be the same as comparisons will be between between map i of proj.curr and map i of proj.fut.}}

\item{proj.fut}{\code{SpatRaster}. One or more ESM/SDM binary projections at future time.}
}
\value{
\code{list} containing:
\itemize{
\item{RangeShift}: \code{SpatRaster}. The range comparisons between current and future (\emph{see Details for the values}).
\item{RangeShift.table}: \code{data.frame}. Summary of the comparisons between maps. containing 9 columns:
\itemize{
\item{layer}. The layer names of proj.fut used to make the comparisons.
\item{Pixel.Absence.Stable}. The number of pixels that are considered as absent for both time period.
\item{Pixel.Presence.Current.Only}. The number of pixels considered as "presence" in proj.curr but not in proj.fut.
\item{Pixel.Presence.Future.Only}. The number of pixels considered as "presence" in proj.fut but not in proj.curr.
\item{Pixel.Presence.Stable}. The number of pixels considered as "presence" both in proj.curr and proj.fut.
\item{Current.Range.size}. The number of pixels considered as "presence" in proj.curr.
\item{Future.Range.size}. The number of pixels considered as "presence" in proj.fut assuming that species is
not limited by their dispersal capacities.
\item{Perc.Loss}. The percentage of pixel considered as "presence" in proj.curr and becoming "absent"
in proj.fut compared to the current range size. 
\item{Perc.Gain}. The percentage of pixel considered as "absent" in proj.curr and becoming 
"present" in proj.fut compared to the current range size. 
}
}
}
\description{
This function compare ESM or SDM binary projections between two time periods by comparing the change 
in suitable/Unsuitable pixels. It also returns percentage of loss and gained
}
\details{
The comparisons betwween SDM/ESM binary predictions are made following this formula: 
\bold{RangeShift = proj.curr + 2*proj.fut}. 
Therefore, when the value is equal to 0. It correspond to a pixel considered as "absent"/"unsuitable" in both 
time period. 
When it is 1, pixels considered as "presence" in proj.curr  but not in proj.fut \bold{(pixel lost)}.
When it is 2, pixels considered as "presence" in proj.fut  but not in proj.curr \bold{(pixel gained)}.
When it is 3, pixels are considered as "present" in both time periods \bold{(pixel Stable)}.
}
\seealso{
\code{\link{ESM_Projection}}, \code{\link{ESM_Ensemble.Projection}}, \code{\link{ESM_Threshold}}, \code{\link{ESM_Binarize}}
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com}
}
