% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ESM_Splachnum.Data}
\alias{ESM_Splachnum.Data}
\title{Species occurrence data to perform ESM}
\format{
A data.frame with 36 lines and 50 columns in the GBIF format:
\describe{
\item{decimalLongitude}{Longitude in the CRS EPSG:4326}
\item{decimalLatitude}{Latitude in the CRS EPSG:4326}
}
}
\usage{
ESM_Splachnum.Data
}
\description{
Contains a filtered dataset obtained from GBIF.org for \emph{Splachnum melanocaulon},
a rare endangered species in Scandinavia.
}
\details{
The data has been downloaded from GBIF.org and afterwards filtered using the default
parameter of the function "clean_coordinates" from the "CoordinateCleaner" R package.
The temporal data is between 1981 and 2024 and coordinate uncertainty is less than 500 m.
}
\examples{
data(ESM_Splachnum.Data)
}
\references{
GBIF.org (16 December 2024) GBIF Occurrence Download. \doi{10.15468/dl.a5fhfv}.
}
\author{
Flavien Collart
}
\keyword{datasets}
