% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_Evaluation.R
\name{ESM_Variable.Contributions}
\alias{ESM_Variable.Contributions}
\title{Variable contribution in ESMs}
\usage{
ESM_Variable.Contributions(ESM.Mod, ESM.ensembleMod)
}
\arguments{
\item{ESM.Mod}{The object returned by \code{\link{ESM_Modeling}}.}

\item{ESM.ensembleMod}{The object returned by \code{\link{ESM_Ensemble.Modeling}}.}
}
\value{
a \code{dataframe} with contribution values (i.e. proportional contribution) for each variable and model
}
\description{
This function evaluates the full model which is used for projections and provides thresholds to produce binary maps.
}
\details{
Calculates the ratio between sum of weights (e.g., AUC, TSS,..) of bivariate models where the variable of interest was used 
and the sum of weights of bivariate models where the variable of interest was not used. The ratio is corrected for the number of models 
with or without the variable of interest. 

This ratio gives an indication on the proportional contribution of the variable in the final ensemble model. A value of higher than 1 
indicates that the focal variable has a higher contribution than average.

In the case of multiple methods (e.g., GLM,...), the contributions are counted per method. For ensemble model, the contributions 
are then weighted means (based on the weighting score as chosen in \code{\link{ESM_Ensemble.Modeling}} of single methods.

For the use of this function, please refer to the manual of ESM_Modeling.
}
\seealso{
\code{\link{ESM_Modeling}}, \code{\link{ESM_Ensemble.Modeling}}, \code{\link{ESM_Projection}}, \code{\link{ESM_Ensemble.Projection}}
}
\author{
Olivier Broennimann \email{Olivier.Broennimann@unil.ch} with contributions of Heidi Mod and Daniel Scherrer 
for the ecospat package and adapted for this package by Flavien Collart \email{flaviencollart@hotmail.com}
}
