\name{GDS} 
\Rdversion{1.1} 
\alias{GDS} 
\title{Descriptive statistics}
\description{Print basic description of a dataset}
\usage{ 
GDS(mat)
}

\arguments{ 
\item{mat}{A matrix giving the performance of each organism on each resource, with organisms as rows} 
}

\value{
	\item{}{Print the number of organisms, resources, interactions, and the fraction of interactions >0.}
}
