% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chelsa.R
\name{get_topography}
\alias{get_topography}
\title{Download topographic data}
\usage{
get_topography(
  var.names,
  res = "1KM",
  aggr = "md",
  path = getwd(),
  extent = NULL,
  mask = FALSE,
  compress = TRUE,
  timeout = 300L
)
}
\arguments{
\item{var.names}{\code{character}. Vector containing the variable names to download. See details for more information.}

\item{res}{\code{character}. Resolution of the data. Available: . \emph{Default: '1KM'}.}

\item{aggr}{\code{character}. Aggregating factor. Available. \emph{Default: 'md'}.}

\item{path}{\code{character}.The path to store the climatic grids.}

\item{extent}{The extent to which you want to crop your climatic datasets. extent must be either a 'SpatRaster', 'SpatVector', or'SpatExtent'.}

\item{mask}{\code{logical}. Do you want to mask your maps? \emph{Note that when TRUE, extent must be a 'SpatRaster' or a 'SpatVector'. }}

\item{compress}{\code{logical}. When extent is provided, do you want to compress the climatic grids. if TRUE, the compression
will be "COMPRESS=DEFLATE", "PREDICTOR=2" and "ZLEVEL=6". \emph{Default: TRUE}.}

\item{timeout}{\code{integer}. The maximum downloading time allowed for each map in seconds. \emph{Default: 300L}.
#' @details  
\describe{}}
}
\value{
"done" when finished
}
\description{
Download and crop topographic variables from Amatulli et al (2017) using GMTED elevation data
}
\examples{
 \donttest{
}
}
\references{
Amatulli
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com}
}
