% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_Loaders.R
\name{Load_ESM}
\alias{Load_ESM}
\alias{Load_ESM_Modelling}
\alias{Load_ESM_Ensemble.Modelling}
\alias{Load_ESM_Projection}
\title{Functions to load the ESM objects}
\usage{
Load_ESM_Modelling(sp.name)

Load_ESM_Ensemble.Modelling(sp.name)

Load_ESM_Projection(sp.name)
}
\arguments{
\item{sp.name}{\code{character}. The species name given in sp.name argument from ESM_Modelling()}
}
\value{
a usable ESM object
}
\description{
These functions allow the user to easily load their ESM objects
}
\details{
The function allow to load the most recent ESM object. Note that the ESM folder 
should be in your working directory
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com}
}
