% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_Evaluation.R
\name{Max_MCC}
\alias{Max_MCC}
\title{Compute the maximum value of Matthew’s Correlation Coefficient}
\usage{
Max_MCC(Pred, Sp.occ)
}
\arguments{
\item{Pred}{\code{numeric}. A vector of predicted probabilities.}

\item{Sp.occ}{\code{numeric}. A vector of speceis observations (1 = presence, 0 = absence)}
}
\value{
a \code{list} containing:
\itemize{
\item{table}. \code{matrix} containing the computed MCC for different threshold.
\item{max.MCC}. \code{numeric}. The maximum value of MCC.
\item{max.threshold}. \code{numeric}. The threshold that maximizes the MCC
}
}
\description{
This function computes the Smooth continuous Boyce Index of Liu et al (2024) using 5 smoothing techniques and make an ensemble.
}
\details{
This function calculates the Matthews Correlation Coefficient (MCC) along different thresholds, between 0.01 and 1 with a 0.01 increment, and return the maximum values.
MCC is a metric particularly useful for imbalanced datasets. Unlike other performance measures, MCC is derived from the contingency matrix and represents the Pearson product-moment 
correlation between observed and predicted classifications. It is unique among binary classification metrics in that it yields a high 
score only when the model accurately predicts both the majority of positive and negative instances. MCC values range from −1 to +1, 
where +1 indicates perfect classification, −1 indicates total misclassification, and 0 corresponds to random guessing, as with a coin toss.
}
\examples{
data(ESM_Species.Env)
MCC <- Max_MCC(
Pred = ESM_Species.Env$ESM_Campylophyllum_halleri,
Sp.occ = ESM_Species.Env$Campylophyllum_halleri
)
MCC$max.MCC

}
\references{
Chicco, D., Jurman, G. The advantages of the Matthews correlation coefficient (MCC) over F1 score and accuracy in binary classification evaluation. BMC Genomics 21, 6 (2020). https://doi.org/10.1186/s12864-019-6413-7
}
\seealso{
\code{\link{ESM_Threshold}}
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com}.
}
