\name{nullweb} 
\Rdversion{1.1} 
\alias{nullweb} 
\title{Null model of binary interactions}
\description{}
\usage{ 
nullweb(ref)
}

\arguments{ 
\item{ref}{An interaction network} 
}

\value{
	\item{}{A null interaction network}
}
 
\details{Based on Bascompte et coll. 2003 PNAS}

\references{}

\examples{}