\name{scale} 
\Rdversion{1.1} 
\alias{scale} 
\title{Scaling of a vector}
\description{Returns a vector scaled between two values}
\usage{ 
scale(v,m=0,M=1)
}

\arguments{ 
\item{v}{Any numerical vector}
\item{m}{The minimal value}
\item{M}{The maximal value}
}

\value{
	\item{}{A vector scaled between m and M}
}