% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_Evaluation.R
\name{Smooth_CBI}
\alias{Smooth_CBI}
\title{Compute the Smooth continuous Boyce Index (SBI) of Liu et al (2024)}
\usage{
Smooth_CBI(
  pres,
  abs,
  ktry = 10,
  method = "all",
  mean.CBI = TRUE,
  cor.method = "spearman"
)
}
\arguments{
\item{pres}{\code{numeric}. Predictions from sites where species is present.}

\item{abs}{\code{numeric}. Predictions from random sites (e.g., background points, pseudo-absences).}

\item{ktry}{\code{integer}. Basis dimension for smoothers. Generally, the default value 10 is ok.}

\item{method}{\code{character}. Either 'all' to compute all methods or vector containing a subset of the 5 available smoothing 
methods to compute the SBI. Available methods:'tp','cr','bs','ps','ad'. \emph{Default: 'all'} See details for more information.}

\item{mean.CBI}{\code{logical}. Should the ensemble model be perform to compute the SBI?}

\item{cor.method}{\code{character}. The method used to compute the correlation. Should be either 'pearson','spearman','kendall'. 
\emph{Default: 'spearman}}
}
\value{
a \code{list} containing:
\itemize{
\item{SBI}. \code{matrix} containing the computed SBI. SBI.m is the value from the ensemble model.
\item{pred.CBI}. \code{matrix}. Predicted values resulted from the smoothing method used to compute the CBI. 
The first column corresponds to the observed predictions.
}
}
\description{
This function computes the Smooth continuous Boyce Index of Liu et al (2024) using 5 smoothing techniques and make an ensemble.
}
\details{
The Boyce index (BI) is usually computed by comparing relative frequencies between probabilities from presence and background sites 
within a series of bins. However, this method is problematic, particularly for rare species, to obtain accurate estimates of the BI.

To correct this issue, Liu et al (2024) proposed the Smooth Boyce Index (SBI), which was shown to be better than regular BI. 
To do so, regression splines, with a binomial family and using mgcv::gam function, are fitted by using the predicted suitability 
from SDMs as the independent variable and species observation (1 = presence, 0 = random points) as the dependent variable. 
The fitted values are then predicted and a correlation between the observed prediction values and the fitted one is realized.
An ensemble of these models (called 'm') can be realized by applying a mean across the predicted prediction values from the different 
techniques. A correlation is then performed between these values and the observed ones. This ensemble can be realized with the argument
mean.CBI = TRUE.

5 smoothing techniques have been tested in Liu et al (2024) and are implemented here:
\itemize{
\item{tp}. Thin plate regression splines. 
\item{cr}. Cubic regression splines.
\item{bs}. B-splines.
\item{ps}. P-splines.
\item{ad}. Adaptive smoothers. 
}
}
\examples{
data(ESM_Species.Env)
SBI <- Smooth_CBI(
pres = ESM_Species.Env$ESM_Campylophyllum_halleri[ESM_Species.Env$Campylophyllum_halleri==1],
abs = ESM_Species.Env$ESM_Campylophyllum_halleri[ESM_Species.Env$Campylophyllum_halleri==0],
)
SBI$SBI

}
\references{
Liu, C., Newell, G., White, M. and Machunter, J. (2024), Improving the estimation of the Boyce index using statistical smoothing methods for evaluating species distribution models with presence-only data. \emph{Ecography}, e07218. \doi{10.1111/ecog.07218}
}
\seealso{
\code{\link{ESM_Modeling}}, \code{\link{ESM_Null.Models}}, \code{\link{ESM_Pooling.Evaluation}}
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com} from the code available in Liu et al (2024).
}
