% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_Tools.R
\name{ESM_Binarize}
\alias{ESM_Binarize}
\title{Binarize probability values using a threshold}
\usage{
ESM_Binarize(proj, thr)
}
\arguments{
\item{proj}{a \code{SpatRaster}, \code{data.frame}, \code{matrix}, or \code{numeric} containing the data to binarize.}

\item{thr}{\code{numeric}. threshold to binarize the probabilities. \bold{must be a single value}.}
}
\value{
An object of the same class as proj: \code{SpatRaster}, \code{data.frame}, \code{matrix}, or \code{numeric}.
}
\description{
This function binarizes probability values based on a specific threshold
}
\details{
\describe{
Probabilities strictly below the threshold will be assigned to 0 (
environmentally unsuitable for the species) while probabilities greater or 
equal to the threshold will be 1 (environementally suitable).
}
}
\examples{
## Generate a vector to binarize
proj <- seq(0,1000, by = 100)
# Binraization of the vector
ESM_Binarize(proj = proj, thr = 400)

}
\seealso{
\code{\link{ESM_Projection}}, \code{\link{ESM_Ensemble.Projection}}, \code{\link{ESM_Threshold}} ,\code{\link{ESM_Range.Shift}}
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com}
}
