% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ESM_Env}
\alias{ESM_Env}
\title{Environmental SpatRaster to perform ESM with the ESM package}
\format{
A SpatRaster with 5 layers:
\describe{
\item{ch_edaphic_eivdescombes_pixel_r}{}
\item{sradY}{Annual solar radiation}
\item{bio3_tiso_8110_LV95}{}
\item{bio15_ps_8110_LV95}{}
}
}
\usage{
ESM_Env
}
\description{
Contains stack of 4 predictors in the Western Swiss Alps
}
\examples{
library(terra)
data(ESM_Env)
ESM_Env <- terra::unwrap(ESM_Env)

}
\references{
Collart, F., Kiebacher, T., Quetsch, M., Broennimann, O., Guisan, A. & Vanderpoorten, A. 2024. To what extent can we predict 
variation of bryophyte and tracheophyte community composition at fine spatial scale along an elevation gradient?
\emph{STOTEN}, 926:171741. \doi{10.1016/j.scitotenv.2024.171741}.
}
\author{
Flavien Collart
}
\keyword{datasets}
