% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ESM_Env}
\alias{ESM_Env}
\title{Environmental SpatRester to perform ESM with the ESM package}
\format{
A SpatRaster with 5 layers:
\describe{
\item{ddeg0}{Growing degree-days above 0C}
\item{mind68}{moisture index for month June to August}
\item{srad68}{solar radiation for month June to August}
\item{slope25}{average of slopes at 25m resolution}
\item{topos25}{average of topographic positions at 25m resolution}
}
}
\usage{
ESM_Env
}
\description{
Contains stack of 5 predictors in the Western Swiss Alps, extracted from the ecospat package
}
\examples{
library(terra)
data(ESM_Env)
ESM_Env <- terra::unwrap(ESM_Env)

}
\author{
Flavien Collart and Olivier Broennimann
}
\keyword{datasets}
