% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_Tools.R
\name{ESM_Generate.ODMAP}
\alias{ESM_Generate.ODMAP}
\title{Generate and prefill ODMAP table}
\usage{
ESM_Generate.ODMAP(ESM.Mod = NULL, ESM.ensembleMod = NULL, ask.to.fill = TRUE)
}
\description{
Generate and prefill ODMAP table
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com}
}
