% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_Models.options.R
\name{ESM_Models.Options}
\alias{ESM_Models.Options}
\title{Model parameters for ESMs}
\usage{
ESM_Models.Options(ANN = NULL, CTA = NULL, GLM = NULL, GBM = NULL)
}
\arguments{
\item{ANN}{a \code{list}. with the objects:
\itemize{
\item{size} an \code{integer}. The numer of units in the hidden layers. \emph{Default}: 8.
\item{decay} a \code{numeric}. The weight decay. \emph{Default}: 0.001.
\item{rang} a \code{numeric}. Initial random weights. \emph{Default}: 0.1.
\item{maxit} an \code{integer}.The maximal number of iterations. \emph{Default}: 200.
}}

\item{CTA}{a \code{list}. with the objects:
\itemize{
\item{na.action} The default action to remove observations with NA in the response variable. \emph{Default}: \code{\link[rpart]{na.rpart}}
\item{method} a \code{character}. Only "class" is available
\item{x} a \code{logical}. Keep a copy of the x matrix. \emph{Default}: FALSE.
\item{y} a \code{logical}. Keep a copy of the y matrix. \emph{Default}: TRUE.
\item{model} a \code{logical}. Keep a copy of the model frame. \emph{Default}: FALSE.
\item{control} a \code{list}. a list of options obtained with \code{\link[rpart]{rpart.control}}. \emph{Default}: $xval = 5, minbucket = 5, minsplit = 5, cp = 0, maxdepth =25. for the other parameters see \code{\link[rpart]{rpart.control}} 
}}

\item{GLM}{a \code{list}. with the objects:
\itemize{
\item{type}: \code{character}. Either "linear", "quadratic" or "cubic". \emph{Default}: "quadratic".
\item{test}: \code{character}. Either "AIC" or "none". Perform or not a step AIC selection for the formula. \emph{Default}: "none".
\item{family}: error distribution family. Only \code{stats}[binomial(link = 'logit')] was tested thus we discourage to change this parameter to another distribution family.
}}

\item{GBM}{a \code{list}. with the objects:
\itemize{
\item{n.trees}: \code{integer}. Number of trees to fit. \emph{Default}: 1000.
\item{interaction.depth}: \code{integer}. The maximum depth of each tree \code{numeric}. Number of trees to fit. \emph{Default}: 4.
\item{n.minobsinnode}: \code{integer}. The minimum number of observations in the terminal nodes of the trees. \emph{Default}: 5.
\item{shrinkage}: \code{numeric}. The learning rate   \emph{Default}: 0.005.
\item{bag.fraction}: \code{numeric}. The fraction of the training set observations randomly selected to propose the next tree in the expansion. \emph{Default}: 0.5.
\item{train.fraction}: \code{numeric}. The first fraction of observations to fit the model. \emph{Default}: 1.
\item{cv.folds}: \code{integer}. Number of cross-validations to perform. \emph{Default}: 3.
\item{verbose}: \code{logical}. print or not the progress. \emph{Default}: FALSE.
\item{n.cores}: \code{logical}. Number of CPU cores to use. \emph{Default}: \code{NULL}.
}}
}
\value{
a \code{list} of parameters for ESM.
}
\description{
Generate a list of model parameters
}
\details{
For the arguments of each modeling technique, please refer to the manual of \code{\link[nnet]{nnet}}, \code{\link[rpart]{rpart}}, \code{\link[stats]{glm}}, and \code{\link[gbm]{gbm}}.
}
\examples{
## Perform a GLM with step AIC to select the best structure and allows linear, quadratic and cubic terms
models.options = ESM_Models.Options(GLM=list(test="AIC",type="cubic"))
## Perform a GLM with linear and quadratic terms but does not perform a selection of the structure
models.options = ESM_Models.Options(GLM=list(test="none",type="quadratic"))

}
\seealso{
[ESM_Modeling], [ESM_Ensemble.Modeling],  [ESM_Projection], [ESM_Ensemble.Projection]
}
\author{
Flavien Collart from the code of biomod2::BIOMOD_ModelingOptions
}
