% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_Projection.R
\name{ESM_Projection}
\alias{ESM_Projection}
\title{Ensemble of Small Models: Projections of Bivariate Models}
\usage{
ESM_Projection(
  ESM.Mod,
  new.env,
  name.env,
  parallel = FALSE,
  n.cores = 1,
  save.obj = T
)
}
\arguments{
\item{ESM.Mod}{The object returned by \code{\link{ESM_Modeling}}.}

\item{new.env}{a \code{data.frame}, \code{matrix} or a \code{SpatRaster} of the predictors where the models 
should be projected. Note that the colnames or the names of the predictors should be exactly the same as the one used in \code{\link{ESM_Modeling}}.}

\item{name.env}{a \code{character} which will be used to generate a folder to save the individual projections.}

\item{parallel}{\code{logical}. Allows or not parallel job using the functions makeCluster.}

\item{n.cores}{\code{integer}. Number of CPU cores used to make the models.}

\item{save.obj}{\code{logical}. Allows or not the final output.}
}
\value{
a \code{list} containing: 
\itemize{
\item{projection.path}: a \code{character} the path to each projection.
\item{model.info}: a \code{list} of the models used, their options, the combination of bivariate models (which.biva),
the modeling ID, the path to the folder where are the stored the models (biva.path), and the failed models.
\item{name.env}: a \code{character} the value given to the argument name.env.
\item{proj.type}: a \code{character} which is either "data.frame" or "SpatRaster" depending on the class of new.env
}
}
\description{
Project bivariate models
}
\references{
Lomba, A., L. Pellissier, C.F. Randin, J. Vicente, F. Moreira, J. Honrado and A. Guisan. 2010. Overcoming the rare species 
modelling paradox: A novel hierarchical framework applied to an Iberian endemic plant. \emph{Biological Conservation}, \bold{143},2647-2657.

Breiner F.T., A. Guisan, A. Bergamini and M.P. Nobis. 2015. Overcoming limitations of modelling rare species by using ensembles of small models. \emph{Methods in Ecology and Evolution}, \bold{6},1210-1218.

Breiner F.T., Nobis M.P., Bergamini A., Guisan A. 2018. Optimizing ensembles of small models for predicting the distribution of species with few occurrences. \emph{Methods in Ecology and Evolution}. \doi{10.1111/2041-210X.12957}
}
\seealso{
[ESM_Modeling], [ESM_Ensemble.Modeling], [ESM_Ensemble.Projection]
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com} based on the previous code written by Frank Breiner 
and Mirko Di Febbraro with the contributions of Olivier Broennimann and Flavien Collart
}
