% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_Tools.R
\name{ESM_Range.Shift.Continuous}
\alias{ESM_Range.Shift.Continuous}
\title{Ensemble of Small Models: Range Shift between projections with continuous predictions}
\usage{
ESM_Range.Shift.Continuous(
  proj.curr,
  proj.fut,
  occ.pres = NULL,
  test.significance = TRUE
)
}
\arguments{
\item{proj.curr}{\code{SpatRaster}. One ESM/SDM continuous projections at current time. 
\emph{Note that if the number of layer is greater than one for proj.curr, the number of layer 
in proj.fut must be the same as comparisons will be between between map i of proj.curr and map i of proj.fut.}}

\item{proj.fut}{\code{SpatRaster}. One or more ESM/SDM continuous projections at future time. \emph{Note that
the map names should be different than in proj.curr}.}

\item{occ.pres}{\code{matrix} or \code{data.frame} containing the X and Y coordinate 
of the species occurrences.\emph{Default: NULL}.}

\item{test.significance}{\code{logical}. Should the change in suitability between 
present and future should be tested? \emph{Default: TRUE}.}
}
\value{
\code{list} containing:
\itemize{
\item{suitability.Change.Map}. \code{SpatRaster}. The difference between present and future time in suitability.
\item{results.All.pixels}. A \code{list} containing:
\itemize{
\item{summary.Suitability}. \code{matrix}. The summary in suitability values of the different maps.
\item{centroids}. \code{matrix}. Centroid coordinates. See details for more information
\item{centroid.shift}. a squared \code{matrix} with the distance in meter between each centroid.
\item{centroid.direction}. \code{matrix}. The direction in degree between the centroid at present time
to the centroid in the future.
\item{significance.SuitabilityDiff}. \code{list}. The result of the different statistical tests comparing 
the suitability values between the two time periods. Only available when test.significance = TRUE.
}
\item{results.on.occurrences}. A \code{list}, only present when occ.pres is not NULL, containing:
\itemize{
\item{summary.Suitability}. \code{matrix}. The summary in suitability values predicted on current 
species occurrences at different time periods.
\item{significance.SuitabilityDiff}. \code{list}. The result of the different statistical tests comparing 
the suitability values extracted on occurrences between the two time periods. Only available when 
test.significance = TRUE.
}
}
}
\description{
This function compare ESM or SDM continuous projections between two time periods by using 
model projections on the whole study area and model predictions on observation data.
}
\details{
This functions compute different metric to infer range change using continuous data:
\enumerate{
\item A Difference in habitat suitability between present and future using all pixels from the study area or using only 
occurrences. If test.significance = TRUE, a friedman test followed by a post-hoc paired wilcox test is performed to 
test if there is a significant difference in suitability between the time periods.
\item A distance and a direction between the centroids defined at present and future time. To do so,
each centroid is computed based on a weighted mean across all pixel centroids of the studied area, 
weighting their x or y values with their respective suitability value following Maclean & Early (2023).
The distance is computed by first projecting centroid coordinates in ESPG:4326 and then computed with 
sp::spDists. The direction is computed using argosfilter::bearing
}
}
\references{
Maclean, I.M.D., Early, R. 2023. Macroclimate data overestimate range shifts of plants 
in response to climate change. \emph{Nat. Clim. Chang}. \bold{13}, 484–490. 
\doi{10.1038/s41558-023-01650-3}.
}
\seealso{
\code{\link{ESM_Projection}}, \code{\link{ESM_Ensemble.Projection}}
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com}
}
