% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ESM_Splachnum.Env}
\alias{ESM_Splachnum.Env}
\title{Environmental SpatRaster to perform ESM in Scandinavia}
\format{
A SpatRaster with 4 layers:
\describe{
\item{bio1}{Annual Mean Temperature averaged between 1981-2010}
\item{bio3}{Isothermalithy. It is equal to: Mean Diurnal 
Range/Temperature Annual Range * 100. The values have been averaged between 
1981 and 2010}
\item{bio5}{Max Temperature of Warmest Month averaged between 1981-2010}
\item{northness}{Northness.close to 1 corresponds to a northern exposition on a 
vertical slope, while a value close to -1 corresponds to a very steep southern slope}
}
}
\usage{
ESM_Splachnum.Env
}
\description{
Contains stack of 4 predictors at ~20km resolution in Scandinavia at present time.
}
\details{
Data were downloaded using get_Chelsa.Clim and get_Topography and aggregated to ~20km.
}
\examples{
library(terra)
data(ESM_Splachnum.Env)
ESM_Splachnum.Env <- terra::unwrap(ESM_Splachnum.Env)

}
\references{
Amatulli, G., Domisch, S., Tuanmu, M.-N., Parmentier, B., Ranipeta, A., Malczyk, J., and Jetz, W. (2018) A suite of global, cross-scale topographic variables for environmental and biodiversity modeling. \emph{Scientific Data}. \bold{5}, 180040. \doi{10.1038/sdata.2018.40}. 

Karger, D.N., Conrad, O., Böhner, J., Kawohl, T., Kreft, H., Soria-Auza, R.W., Zimmermann, N.E., Linder, P., Kessler, M. (2017). Climatologies at high resolution for the Earth land surface areas. \emph{Scientific Data}. \bold{4}, 170122. \doi{10.1038/sdata.2017.122}.
}
\author{
Flavien Collart
}
\keyword{datasets}
