% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_Evaluation.R
\name{ESM_Variable.Contributions}
\alias{ESM_Variable.Contributions}
\title{Variable contribution in ESMs}
\usage{
ESM_Variable.Contributions(ESM.Mod, ESM.ensembleMod)
}
\arguments{
\item{ESM.Mod}{The object returned by \code{\link{ESM_Modeling}}.}

\item{ESM.ensembleMod}{The object returned by \code{\link{ESM_Ensemble.Modeling}}.}
}
\value{
a \code{dataframe} with contribution values (i.e. proportional contribution) by variable and model
}
\description{
This function evaluates the full model which is used for projections and provides thresholds to produce binary maps.
}
\details{
Calculates the ratio between sum of weights of bivariate models where a focal variable was used and sum of weights of bivariate 
models where the focal variable was not used. The ratio is corrected for the number of models with or without the focal variable. 
This ratio gives an indication on the proportional contribution of the variable in the final ensemble model. A value of higher than 1 
indicate that the focal variable has a higher contribution than average.
In the case of multiple methods (e.g., GLM,...), the contributions are counted per method. For ensemble model, the contributions 
are then weighted means (based on the weighting score as chosen in ecospat.ESM.EnsembleModeling of single methods.
}
\seealso{
[ESM_Modeling], [ESM_Ensemble.Modeling],  [ESM_Projection], [ESM_Ensemble.Projection]
}
\author{
Olivier Broennimann \email{Olivier.Broennimann@unil.ch} with contributions of Heidi Mod and Daniel Scherrer 
and the adaptations of Flavien Collart \email{flaviencollart@hotmail.com}
}
