% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_Chelsa.Clim}
\alias{get_Chelsa.Clim}
\title{Download chelsa data}
\usage{
get_Chelsa.Clim(
  var.names,
  time = "1981-2010",
  gcm = NULL,
  ssp = NULL,
  path = getwd(),
  extent = NULL,
  mask = FALSE,
  compress = TRUE,
  timeout = 300L
)
}
\arguments{
\item{var.names}{\code{character}. Vector containing the variable names to download. See details for more information.}

\item{time}{\code{character}. The time period(s) wanted. Must be: '1981-2010','2011-2040','2041-2070', and/or'2071-2100'.
\emph{Default: '1981-2010'}}

\item{gcm}{\code{character}. A vector containing the Global Circulation Models for which you want to download the datasets. 
Only needed when future time periods are provided. Must be: 'gfdl-esm4','ipsl-cm6a-lr','mpi-esm1-2-hr','mri-esm2-0' and/or,'ukesm1-0-ll'.
\emph{Default: NULL}.}

\item{ssp}{\code{character}. A vector containing the Shared Socio-economic Pathways for which you want to download the datasets. 
Only needed when future time periods are provided. Must be: '126', '370' and/or '585'.\emph{Default: NULL}.}

\item{path}{\code{character}.The path to store the climatic grids.}

\item{extent}{The extent to which you want to crop your climatic datasets. extent must be either a 'SpatRaster', 'SpatVector', or'SpatExtent'.}

\item{mask}{\code{logical}. Do you want to mask your maps? \emph{Note that when TRUE, extent must be a 'SpatRaster' or a 'SpatVector'.}}

\item{compress}{\code{logical}. When extent is provided, do you want to compress the climatic grids? if TRUE, the compression
will be: "COMPRESS=DEFLATE", "PREDICTOR=2" and "ZLEVEL=6". \emph{Default: TRUE}.}

\item{timeout}{\code{integer}. The maximum downloading time allowed for each map in seconds. \emph{Default: 300L}.}
}
\value{
\code{character}. A vector containing the path to the downloaded files.
}
\description{
Download and crop climatic variables at different time period from CHELSA v.2.1.
}
\details{
\describe{
get_chesla download and crop climatic variables at different time periode from CHELSA v.2.1.
Some variables are only available at the present time (var.pres) while other are also available in the futur (var.full).
The full variables are to be combined with a future scenario by choosing a global circular model (gcm.chelsa) and a shared-socioeconomic pathway (ssp.chelsa)
The description of the different variables is available at the point 7 (p.11) of the following link :
\link{https://chelsa-climate.org/wp-admin/download-page/CHELSA_tech_specification_V2.pdf}
}
}
\examples{
 \donttest{
# get_Chelsa.Clim("bio1")
}
}
\references{
Karger, D.N., Conrad, O., Böhner, J., Kawohl, T., Kreft, H., Soria-Auza, R.W., Zimmermann, N.E., Linder, P., Kessler, M. (2017). Climatologies at high resolution for the Earth land surface areas. \emph{Scientific Data}. \bold{4}, 170122. \doi{10.1038/sdata.2017.122}.

Karger D.N., Conrad, O., Böhner, J., Kawohl, T., Kreft, H., Soria-Auza, R.W., Zimmermann, N.E, Linder, H.P., Kessler, M. (2018): Data from: Climatologies at high resolution for the earth’s land surface areas. \emph{EnviDat}. \doi{10.16904/envidat.228.v2.1}.
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com} with contributions of Adèle Hotermans
}
