% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chelsa.R
\name{get_topography}
\alias{get_topography}
\title{Download topographic data}
\usage{
get_topography(
  var.names,
  res = "1KM",
  aggr = "md",
  path = getwd(),
  extent = NULL,
  mask = FALSE,
  compress = TRUE,
  timeout = 300L
)
}
\arguments{
\item{var.names}{\code{character}. Vector containing the variable name(s) to download. Can be either 'all' to download all variables or a subset of the variables
avaible in Amatulli et al (2017). See details for more information.}

\item{res}{\code{character}. Resolution of the data. One element from: '1KM','5KM','10KM','50KM','100KM'. \emph{Default: '1KM'}.}

\item{aggr}{\code{character}. Vector containing the aggregating factor(s). Can be either 'all' to download all the aggregating factors or a subset 
of c('md','mn','mi','ma','sd'). \emph{Default: 'md'}. See details for more information.}

\item{path}{\code{character}.The path to store the climatic grids.}

\item{extent}{The extent to which you want to crop your climatic datasets. extent must be either a 'SpatRaster', 'SpatVector', or'SpatExtent'.}

\item{mask}{\code{logical}. Do you want to mask your maps? \emph{Note that when TRUE, extent must be a 'SpatRaster' or a 'SpatVector'. }}

\item{compress}{\code{logical}. When extent is provided, do you want to compress the climatic grids. if TRUE, the compression
will be "COMPRESS=DEFLATE", "PREDICTOR=2" and "ZLEVEL=6". \emph{Default: TRUE}.}

\item{timeout}{\code{integer}. The maximum downloading time allowed for each map in seconds. \emph{Default: 300L}.
#' @details  
\describe{}}
}
\value{
\code{character}. A vector containing the path to the downloaded files.
}
\description{
Download and crop topographic variables from Amatulli et al (2017) using GMTED elevation data
}
\examples{
 \donttest{
# get_topography("elevation")
}
}
\references{
Amatulli, G., Domisch, S., Tuanmu, M.-N., Parmentier, B., Ranipeta, A., Malczyk, J., and Jetz, W. (2018) A suite of global, cross-scale topographic variables for environmental and biodiversity modeling. \emph{Scientific Data}. \bold{5}, 180040. \doi{10.1038/sdata.2018.40}.
}
\author{
Flavien Collart \email{flaviencollart@hotmail.com}
}
